refactor(landing): remove landing- prefix from section filenames
